requirements.txt
asgiref
crispy-bootstrap5
distlib
dj-database-url
Django
django-crispy-forms
django-email
django-environ
django-filter
django-humanize
djangorestframework
filelock
humanize
Markdown
numpy
opencv-contrib-python
pipenv
platformdirs
psycopg2
psycopg2-binary
pytz
sqlparse
typing_extensions
tzdata
